Refactor: BaseCertification 의존성 제거 #100
